bot_code/requirements.txt
nextcord==2.0.0a8
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
python-dateutil==2.8.0
python-dotenv==0.15.0
fastDamerauLevenshtein==1.0.7
beautifulsoup4==4.11.1
pyenchant==3.2.2